+++ src/plugins/ordering/index.ts
@@ -1,3 +1,4 @@
+export * from './misc/services';
 export * from './no/noOrdering.component';
 export * from './no/noOrdering.interface';
 export * from './single/singleOrdering.component';